Renamed component -> features and container -> pages
