+++ src/features/NewsCategories/Entertainment/Entertainment.tsx
