+++ src/features/NewsCategories/Technology/Technology.tsx
