+++ src/pages/NewsPage/types.d.ts
@@ -1,4 +1,4 @@
-type NewsContainerProps = {
+type NewsPageProps = {
   news: Object<string>
   category: string
   userInput: string
